move userContext under /components + Fix edge cases
